// src/lib/api/types/listSectorsResponseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */
import type { SectorResponse } from './sectorResponse';

export interface ListSectorsResponseBody {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /**
   * Sector list
   * @nullable
   */
  data: SectorResponse[] | null;
  /** Current page */
  page: number;
  /** Items per page */
  pageSize: number;
  /** Total count */
  total: number;
  /** Total pages */
  totalPages: number;
}